Improving classifier.
fnn_classifier is improved, instead of building the network with
build_network I custme design it with tw hidden layers.

I am not using the percent error, takes too much and can't trust the
results, instead I am doing the activation and getting the class in a
winner takes it all fashion.

SequenceClassifer uses the sequence classification data set, use
recurrent network and using LSTMLayer. It takes a moving
window of size 5 and trains the net over it.

Saving the best executing network (bother for fn and seq_rnn) in the
results folder. Also saving the percent erros.